feat(schd): returning list of preferred nodes for scheduling
Scheduler should return list of preferred nodes for scheduling.
It is up to the user of this libary which can pick first node in
the preferred list or can try on all the nodes in the order returned
by the scheduler.

This can solve the issue where csi controller is not able to create
the volume on first node, it can pick the next one in the list.

Signed-off-by: Pawan <pawan@mayadata.io>